+++ src/pages/AllTwoshotsPage.tsx
@@ -1,0 +1,11 @@
+import { Heading } from "@chakra-ui/react";
+import { Link } from "react-router-dom";
+
+const AllTwoshotsPage = () => (
+  <>
+    <Heading>ALL TWOSHOTs</Heading>
+    <Link to="/">トップページ</Link>
+  </>
+);
+
+export default AllTwoshotsPage;
